### 2026-05-31 #773 @BoellenruecherM
Pre-submission enquiry: dyadicMarkov
Submitting Author Name: Mattia Böllenrücher
Submitting Author Github Handle: @BoellenruecherM
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/BoellenruecherM/dyadicMarkov-public
Submission type: Pre-submission

---

* Paste the full DESCRIPTION file inside a code block below:

```r
Package: dyadicMarkov
Title: Pattern Identification for Dyadic Sequences Using Transition Matrices
Version: 0.1.1
    person(
      given = "Mattia",
      family = "Böllenrücher",
      email = "mattia.boellenruecher@student.unisg.ch"
    ),
    person(
      given = "Mégane",
      family = "Bollenrucher",
      role = "aut"
    ),
    person(
      given = "Jean-Philippe",
      family = "Antonietti",
      role = "aut"
    )
  )
Description: Provides methods for analyzing dyadic interaction sequences using
    transition matrices within the Actor-Partner Interdependence Model (APIM).
    The package supports the computation of empirical transition counts,
    maximum likelihood estimation of transition probabilities and identification
    of interaction patterns in univariate and bivariate dyadic interaction
    sequences.
    stats
```

* Please indicate which category or categories from our [[package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories)](https://ropensci.github.io/dev_guide/policies.html#package-categories) or [[statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories)](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

* Explain how and why the package falls under these categories (briefly, 1-2 sentences).  Please note any areas you are unsure of:

`dyadicMarkov` analyzes categorical dyadic interaction sequences through transition matrices, empirical transition counts, transition-probability estimation, likelihood-ratio tests, and AIC-based comparisons of constrained transition structures. I selected **Time Series Analyses** because the package works with ordered dyadic sequences and Markov-type transition structures, but I am unsure whether this is the best rOpenSci statistical category because the method is specifically focused on dyadic APIM-based interaction patterns rather than general time-series forecasting.

* If submitting a statistical package, have you already [incorporated documentation of standards into your code via the **srr** package](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr)?

Yes. The package includes `@srrstats` and `@srrstatsNA` tags in source and test files. Running

```r
srr::srr_stats_pre_submit(path = ".", quiet = FALSE)
```

reports that all applicable standards have been documented. In the current review version, 36 standards are documented as complied with and 32 standards are documented as not applicable.

`pkgcheck::pkgcheck(".", use_cache = FALSE)` reports positive package status and identifies the package as statistical software, but it currently assigns the package to 0 statistical categories. This is one of the main reasons for opening a pre-submission enquiry before a formal submission.

* Who is the target audience and what are scientific applications of this package?

The target audience is researchers working with categorical dyadic interaction sequences, especially in psychology, behavioral science, communication research, and related fields where dyadic processes are observed over time. The package supports empirical transition counting, transition-probability estimation, and identification of actor, partner, actor-partner, independence, partial bivariate, and complete bivariate interaction patterns in dyadic categorical sequences.

* Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [[our criteria for best-in-category](https://ropensci.github.io/dev_guide/policies.html#overlap)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

There are related R packages for sequence analysis and Markov-chain modeling, such as packages focused on general sequence representation, Markov-chain estimation, or hidden Markov models. However, `dyadicMarkov` addresses a more specific methodological need: dyadic categorical interaction sequences represented through actor-partner transition matrices, with pattern identification linked to APIM-based dyadic interaction structures. The package is therefore not intended to replace general sequence-analysis or Markov-chain packages, but to provide a focused implementation of this specific dyadic transition-matrix methodology.

* (If applicable) Does your package comply with our [guidance around *Ethics, Data Privacy and Human Subjects Research*](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

The package does not collect, store, transmit, or retrieve personal data. It provides statistical methods for user-supplied categorical dyadic sequences. The included example datasets are synthetic package examples. Users may apply the package to human-subjects data, but ethical approval, informed consent, privacy protection, and appropriate data handling remain the responsibility of the user and their institution.

* Any other questions or issues we should be aware of?:

The current development version for review is `dyadicMarkov` 0.1.1. The package is currently on CRAN in an earlier version, but the current review version has not yet been submitted to CRAN. The current review version is available in the public repository and has passed local `R CMD check --as-cran`, built-tarball checks, Win-builder checks on R-release/R-oldrelease/R-devel, GitHub Actions, `srr::srr_stats_pre_submit()`, and `pkgcheck::pkgcheck()`.

My main questions are:

1. Is `dyadicMarkov` in scope for rOpenSci software/statistical software review?
2. Is **Time Series Analyses** the most appropriate statistical category, or should another category be used?
3. Should the package proceed through statistical software review, ordinary software review, or another route?
4. Are there any scope or readiness issues that should be addressed before a formal submission?

## Use of Generative AI

* [x] Generative AI tools were used to produce some of the material in this submission.

Generative AI was used to help draft and polish parts of this pre-submission enquiry text and to organize the checklist of package checks. The package code, documentation, tests, check results, and repository contents were reviewed, edited, and executed by the submitting author. The generated text was manually revised before submission.
