+++ tests/testthat/test-countEmp-deterministic.R
@@ -15,5 +15,10 @@ test_that("countEmp returns exact counts on a tiny hand-constructed chain", {
     nrow = 4L, ncol = 2L, byrow = TRUE
   )
 
-  expect_identical(unclass(got), exp)
+  expect_identical(unname(unclass(got)), exp)
+  expect_identical(rownames(got), c(
+    "focal_1_partner_1", "focal_1_partner_2",
+    "focal_2_partner_1", "focal_2_partner_2"
+  ))
+  expect_identical(colnames(got), c("next_1", "next_2"))
 })
