+++ reference/index.md
@@ -1,0 +1,31 @@
+# Package index
+
+## Empirical transition matrices
+
+- [`countEmp()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmp.md)
+  : Empirical transition counts for univariate dyadic sequences
+- [`countEmpBivariate()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmpBivariate.md)
+  : Empirical transition counts for bivariate dyadic sequences
+
+## Transition probability estimation
+
+- [`mleEstimation()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/mleEstimation.md)
+  : Maximum likelihood estimation of transition probabilities
+
+## Pattern identification
+
+- [`univariatePattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/univariatePattern.md)
+  : Univariate pattern identification for dyadic Markov chains
+- [`bivariateCase()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/bivariateCase.md)
+  : Bivariate case identification for dyadic Markov chains
+- [`partialPattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/partialPattern.md)
+  : Partial bivariate pattern identification by AIC
+- [`completePattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/completePattern.md)
+  : Complete bivariate pattern identification by AIC
+
+## Example data
+
+- [`dyadic_univariate_example`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/dyadic_univariate_example.md)
+  : Synthetic univariate dyadic sequence example
+- [`dyadic_bivariate_example`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/dyadic_bivariate_example.md)
+  : Synthetic bivariate dyadic sequence example
