+++ articles/univariate-workflow.md
@@ -1,0 +1,229 @@
+# Univariate dyadic workflow
+
+## Overview
+
+This vignette shows the univariate workflow implemented in
+`dyadicMarkov`. In the univariate setting, one categorical variable is
+observed repeatedly for the two members of a dyad. The workflow follows
+the single-case method described in Bollenrücher et al. (2023):
+empirical transition counts are computed from the dyadic sequence,
+transition probabilities are estimated by maximum likelihood, and
+restricted actor/partner structures are compared to identify the pattern
+of interaction.
+
+The example uses the data set `dyadic_univariate_example` included in
+the package. The data are synthetic and are used only to illustrate the
+required input structure and the package workflow.
+
+## Data
+
+The univariate example contains one categorical variable for the first
+member (`FM`) and the second member (`SM`) of a dyad. Each row
+corresponds to one measurement occasion. In the first analysis, `FM` is
+the focal sequence and `SM` is the partner sequence.
+
+``` r
+
+utils::data("dyadic_univariate_example", package = "dyadicMarkov")
+
+head(dyadic_univariate_example)
+#>   time FM SM
+#> 1    1  2  1
+#> 2    2  2  1
+#> 3    3  1  2
+#> 4    4  1  1
+#> 5    5  1  1
+#> 6    6  2  1
+dim(dyadic_univariate_example)
+#> [1] 90  3
+```
+
+The states are integer-coded. In this example, both members can take the
+values 1 or 2.
+
+``` r
+
+table(dyadic_univariate_example$FM)
+#> 
+#>  1  2 
+#> 50 40
+table(dyadic_univariate_example$SM)
+#> 
+#>  1  2 
+#> 47 43
+```
+
+## Empirical transition counts
+
+The first step is to compute empirical transition counts with
+[`countEmp()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmp.md).
+For `states = 2`, the rows represent the four possible previous dyadic
+states `(FM_t, SM_t)`, and the columns represent the state of the first
+member at the next time point, `FM_{t+1}`.
+
+``` r
+
+emp_uni <- dyadicMarkov::countEmp(
+  chainFM = dyadic_univariate_example$FM,
+  chainSM = dyadic_univariate_example$SM,
+  states = 2L
+)
+
+emp_uni
+#>                   next_1 next_2
+#> focal_1_partner_1 26     4     
+#> focal_1_partner_2 5      14    
+#> focal_2_partner_1 12     5     
+#> focal_2_partner_2 7      16
+class(emp_uni)
+#> [1] "dyadic_counts" "matrix"        "array"
+```
+
+The resulting object keeps ordinary matrix behavior.
+
+``` r
+
+dim(emp_uni)
+#> [1] 4 2
+rowSums(emp_uni)
+#> focal_1_partner_1 focal_1_partner_2 focal_2_partner_1 focal_2_partner_2 
+#>                30                19                17                23
+```
+
+## Maximum-likelihood transition probabilities
+
+The empirical counts can be converted into transition probabilities
+using
+[`mleEstimation()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/mleEstimation.md).
+Rows with positive totals are normalized independently. For an
+unobserved previous-state combination, the transition probabilities are
+unidentified; the package returns a uniform row as an implementation
+convention.
+
+``` r
+
+fit_uni <- dyadicMarkov::mleEstimation(emp_uni)
+
+round(fit_uni, 3)
+#>                   next_1 next_2
+#> focal_1_partner_1 0.867  0.133 
[…]
