+++ articles/dyadicMarkov-introduction.md
@@ -1,0 +1,133 @@
+# Introduction to dyadicMarkov
+
+## Purpose of the package
+
+`dyadicMarkov` implements an R workflow for identifying patterns of
+interaction in categorical dyadic sequences using transition matrices.
+The package is designed for situations in which one or two categorical
+variables are observed repeatedly for both members of one dyad, so that
+the analysis accounts for both temporal dependence and dyadic
+dependence.
+
+The methodological background is developed in a sequence of papers on
+dyadic pattern analysis with the Longitudinal Actor-Partner
+Interdependence Model (L-APIM) and Markov chains. The univariate
+single-case method is described in Bollenrücher et al. (2023), the
+extension to visualization and clustering of similar dyadic behaviors is
+described in Bollenrücher et al. (2024), and the bivariate single-case
+method is described in Böllenrücher et al. (in press). The visualization
+and clustering methodology of Bollenrücher et al. (2024) is
+methodological background; it is not currently part of the exported
+package API.
+
+## Data structure
+
+The package works with categorical dyadic sequences. In the univariate
+case, one categorical variable is observed over time for two members of
+a dyad. For each function call, the first member is the focal individual
+whose next state is modeled, and the second member supplies the partner
+sequence. The roles can be reversed to analyze the other member.
+
+In the bivariate case, two categorical variables are observed over time
+for both members of the dyad. The current implementation of the
+bivariate workflow supports binary variables (`states = 2`). This leads
+to a bivariate empirical count matrix with 16 rows and 2 columns: the 16
+rows represent the four binary lagged components: focal member on the
+main variable, partner member on the main variable, focal member on the
+second variable and partner member on the second variable. The 2 columns
+represent the possible next states of the focal member on the main
+variable.
+
+## Estimation and identification
+
+The package separates estimation from identification. Estimation
+summarizes the observed sequences as empirical transition counts and
+maximum-likelihood transition probabilities. Identification compares the
+observed transition structure with restricted transition structures
+corresponding to interpretable patterns of interaction.
+
+In the univariate workflow, the relevant patterns are actor-partner,
+actor-only, partner-only and independence. In the bivariate workflow,
+the analysis first identifies the global case as trivial, univariate,
+partial bivariate or complete bivariate. A trivial case has no
+subsequent local pattern. A univariate case is followed by
+[`univariatePattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/univariatePattern.md)
+on the main-variable sequences. Partial and complete cases are followed
+by
+[`partialPattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/partialPattern.md)
+and
+[`completePattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/completePattern.md),
+respectively.
+
+## Exported functions
+
+The user-facing workflow is organized around seven exported functions:
+
+- [`countEmp()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmp.md)
+  computes empirical transition counts for one focal sequence in a
+  univariate dyadic sequence.
+- [`mleEstimation()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/mleEstimation.md)
+  estimates transition probabilities from empirical count matrices.
[…]
