+++ LICENSE-text.md
@@ -1,0 +1,4 @@
+# License
+
+    YEAR: 2026
+    COPYRIGHT HOLDER: Mattia Böllenrücher

[…]
