### 2026-06-20 48869f4
Add clean pkgdown homepage
