+++ pkgdown/index.md
@@ -1,0 +1,32 @@
+\# dyadicMarkov
+
+
+
+\## Overview
+
+
+
+`dyadicMarkov` provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov-chain framework.
+
+
+
+The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
+
+
+
+It is designed for single-case dyadic sequence analysis, where two members of a dyad are observed repeatedly over time. The workflow helps researchers describe the temporal dynamics of each member while accounting for the possible influence of their partner.
+
+
+
+\## Installation
+
+
+
+You can install the released version of `dyadicMarkov` from CRAN:
+
+
+
+```r
+
+install.packages("dyadicMarkov")
+
