### 2026-06-20 1082d8f
Use clean pkgdown homepage
