### 2026-06-20 c6896f0
Correct bivariate article citation on website
