+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   bivariate-workflow: bivariate-workflow.html
   dyadicMarkov-introduction: dyadicMarkov-introduction.html
   univariate-workflow: univariate-workflow.html
-last_built: 2026-06-20T11:30Z
+last_built: 2026-06-20T11:40Z
 urls:
   reference: https://boellenruecherm.github.io/dyadicMarkov-public/reference
   article: https://boellenruecherm.github.io/dyadicMarkov-public/articles
