### 2026-06-20 27bea0d
Add full MIT license for GitHub
