### 2026-06-20 edb16ab
Enhance _pkgdown.yml with R-universe sidebar
Added home section with sidebar components for R-universe.
