### 2026-06-20 d74ba30
Remove CRAN page link from pkgdown index
Removed the link to the official CRAN page from the index.
