+++ DESCRIPTION
@@ -3,9 +3,11 @@ Type: Package
 Title: Pattern Identification for Dyadic Sequences Using Transition Matrices
 Version: 0.1.1
 Authors@R: c(
-  person("Mattia", "Böllenrücher", email = "mattia.boellenruecher@student.unisg.ch", role = c("aut", "cre", "cph")),
+  person("Mattia", "Böllenrücher", email = "mattia.boellenruecher@student.unisg.ch", role = c("aut", "cre", "cph"),
+         comment = c(ORCID = "0009-0004-4149-4745")),
   person("Mégane", "Bollenrücher", role = "aut"),
-  person("Jean-Philippe", "Antonietti", role = "aut"))
+  person("Jean-Philippe", "Antonietti", role = "aut",
+         comment = c(ORCID = "0000-0003-0117-4769")))
 Description: Provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
 License: MIT + file LICENSE
 URL: https://boellenruecherm.github.io/dyadicMarkov-public/, https://github.com/BoellenruecherM/dyadicMarkov-public
