### 2026-06-20 9a14ab9
Add authors section to pkgdown configuration
