### 2026-06-20 05f30a8
Correct HTML entity in pkgdown index
Fixed HTML entity for hyphen in package overview.
