+++ pkgdown/index.md
@@ -2,7 +2,7 @@
 
 ## Overview
 
-`dyadicMarkov` provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
+`dyadicMarkov` provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model (L&#8209;APIM) and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
 
 The package is designed for single-case dyadic sequence analysis, where two members of a dyad are observed repeatedly over time. It allows researchers to describe the temporal dynamics of each member while accounting for the possible influence of their partner.
 
