+++ DESCRIPTION
@@ -5,7 +5,8 @@ Version: 0.1.1
 Authors@R: c(
   person("Mattia", "Böllenrücher", email = "mattia.boellenruecher@student.unisg.ch", role = c("aut", "cre", "cph"),
          comment = c(ORCID = "0009-0004-4149-4745")),
-  person("Mégane", "Bollenrücher", role = "aut"),
+  person("Mégane", "Bollenrücher", role = "aut",
+         comment = c(ORCID = "0000-0002-9035-8799")),
   person("Jean-Philippe", "Antonietti", role = "aut",
          comment = c(ORCID = "0000-0003-0117-4769")))
 Description: Provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
