### 2026-06-20 c1507d2
Rearrange URLs in DESCRIPTION file
Updated the URL order in the DESCRIPTION file for clarity.
