+++ authors.md
@@ -19,12 +19,12 @@ Source:
 Böllenrücher M, Bollenrücher M, Antonietti J (2026). *dyadicMarkov:
 Pattern Identification for Dyadic Sequences Using Transition Matrices*.
 R package version 0.1.1,
-<https://boellenruecherm.github.io/dyadicMarkov-public/>.
+<https://github.com/BoellenruecherM/dyadicMarkov-public>.
 
     @Manual{,
       title = {dyadicMarkov: Pattern Identification for Dyadic Sequences Using Transition Matrices},
       author = {Mattia Böllenrücher and Mégane Bollenrücher and Jean-Philippe Antonietti},
       year = {2026},
       note = {R package version 0.1.1},
-      url = {https://boellenruecherm.github.io/dyadicMarkov-public/},
+      url = {https://github.com/BoellenruecherM/dyadicMarkov-public},
     }
