### 2026-06-20 6caa641
Update installation instructions in index.md
Added instructions for installing the development version from R-universe and GitHub.
