### 2026-06-20 3f57055
Add community section to pkgdown sidebar
