### 2026-06-20 05e89b9
Expand contributing guidelines
