### 2026-06-20 767304c
Include rOpenSci Code of Conduct in index.md
Added a note about the rOpenSci Code of Conduct for contributors.
