### 2026-06-20 1ff1ab5
Revise Code of Conduct to link rOpenSci guidelines
Updated the Code of Conduct section to reference the rOpenSci Code of Conduct.
