+++ llms.txt
@@ -79,6 +79,12 @@ Accepted Manuscript\]. Quantitative and Computational Methods in
 Behavioral Sciences. DOI:
 [10.23668/psycharchives.22174](https://doi.org/10.23668/psycharchives.22174)
 
+------------------------------------------------------------------------
+
+Please note that this project follows the [rOpenSci Code of
+Conduct](https://ropensci.org/code-of-conduct/). By contributing to this
+project, you agree to abide by its terms.
+
 # Package index
 
 ## Empirical transition matrices
