+++ index.md
@@ -78,3 +78,9 @@ interdependence model and Markov chains for single-case* \[Author
 Accepted Manuscript\]. Quantitative and Computational Methods in
 Behavioral Sciences. DOI:
 [10.23668/psycharchives.22174](https://doi.org/10.23668/psycharchives.22174)
+
+------------------------------------------------------------------------
+
+Please note that this project follows the [rOpenSci Code of
+Conduct](https://ropensci.org/code-of-conduct/). By contributing to this
+project, you agree to abide by its terms.
