+++ CONTRIBUTING.md
@@ -79,6 +79,8 @@ easier to review and accept.
 
 ## Code of conduct
 
-Please note that the `dyadicMarkov` project is released with a
-Contributor Code of Conduct. By contributing to this project, you agree
-to abide by its terms.
+## Code of conduct
+
+Please note that this project follows the [rOpenSci Code of
+Conduct](https://ropensci.org/code-of-conduct/). By contributing to this
+project, you agree to abide by its terms.
