### 2026-06-20 13c48f7
Polish contributing guidelines
