### 2026-06-20 ba04d96
Regenerate package documentation with ORCIDs
