### 2026-06-20 39b4f28
Sync CRAN comments and documentation polish
