### 2026-06-20 f9e3ebb
Regenerate bivariateCase documentation
