### 2026-06-21 333c9c5
Align member nomenclature with package source
