+++ vignettes/dyadicMarkov-introduction.Rmd
@@ -1,6 +1,5 @@
 ---
 title: "Introduction to dyadicMarkov"
-bibliography: references.bib
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Introduction to dyadicMarkov}
@@ -16,13 +15,13 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 `dyadicMarkov` implements an R workflow for identifying patterns of interaction in categorical dyadic sequences using transition matrices. The package is designed for situations in which one or two categorical variables are observed repeatedly for both members of one dyad, so that the analysis accounts for both temporal dependence and dyadic dependence.
 
-The methodological background is developed in a sequence of papers on dyadic pattern analysis with the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov chains. The univariate single-case method is described in @bollen2023, the extension to visualization and clustering of similar dyadic behaviors is described in @bollen2024, and the bivariate single-case method is described in @bollen2026. The visualization and clustering methodology of @bollen2024 is methodological background; it is not currently part of the exported package API.
+`dyadicMarkov` is based on three methodological papers on dyadic pattern analysis with the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov chains. The univariate single-case method is described by Bollenrücher, Darwiche, and Antonietti (2023). The extension to visualization and clustering of similar dyadic behaviors is described by Bollenrücher, Darwiche, and Antonietti (2024). The bivariate single-case method is described by Böllenrücher, Darwiche, and Antonietti (in press). The visualization and clustering methodology of Bollenrücher, Darwiche, and Antonietti (2024) is methodological background; it is not currently part of the exported package API.
 
 ## Data structure
 
-The package works with categorical dyadic sequences. In the univariate case, one categorical variable is observed over time for two members of a dyad. For each function call, the first member is the focal individual whose next state is modeled, and the second member supplies the partner sequence. The roles can be reversed to analyze the other member.
+The package works with categorical dyadic sequences. In the univariate case, one categorical variable is observed over time for two members of a dyad. For each function call, the first member is the member whose next state is modeled, and the second member supplies the partner sequence. The roles can be reversed to analyze the other member.
 
-In the bivariate case, two categorical variables are observed over time for both members of the dyad. The current implementation of the bivariate workflow supports binary variables (`states = 2`). This leads to a bivariate empirical count matrix with 16 rows and 2 columns: the 16 rows represent the four binary lagged components: focal member on the main variable, partner member on the main variable, focal member on the second variable and partner member on the second variable. The 2 columns represent the possible next states of the focal member on the main variable.
+In the bivariate case, two categorical variables are observed over time for both members of the dyad. The current implementation of the bivariate workflow supports binary variables (`states = 2`). This leads to a bivariate empirical count matrix with 16 rows and 2 columns: the 16 rows represent the four binary lagged components: the first member on the main variable, the second member on the main variable, the first member on the second variable and the second member on the second variable. The 2 columns represent the possible next states of the first member on the main variable.
 
 ## Estimation and identification
 
@@ -34,17 +33,17 @@ In the univariate workflow, the relevant patterns are actor-partner, actor-only,
 
 The user-facing workflow is organized around seven exported functions:
 
-- `countEmp()` computes empirical transition counts for one focal sequence in a univariate dyadic sequence.
[…]
