+++ reference/mleEstimation.md
@@ -37,9 +37,9 @@ chainFM <- c(1L, 2L, 1L, 2L, 2L, 1L)
 chainSM <- c(2L, 1L, 2L, 1L, 1L, 2L)
 emp <- countEmp(chainFM, chainSM, states = 2L)
 mleEstimation(emp)
-#>                   next_1 next_2
-#> focal_1_partner_1 0.500  0.500 
-#> focal_1_partner_2 0.000  1.000 
-#> focal_2_partner_1 0.667  0.333 
-#> focal_2_partner_2 0.500  0.500 
+#>         next_1 next_2
+#> FM1_SM1 0.500  0.500 
+#> FM1_SM2 0.000  1.000 
+#> FM2_SM1 0.667  0.333 
+#> FM2_SM2 0.500  0.500 
 ```
