### 2026-06-21 27b8973
Polish CRAN metadata and workflow terminology
