+++ vignettes/univariate-workflow.Rmd
@@ -20,7 +20,7 @@ The example uses the data set `dyadic_univariate_example` included in the packag
 
 ## Data
 
-The univariate example contains one categorical variable for the first member (`FM`) and the second member (`SM`) of a dyad. Each row corresponds to one measurement occasion. In the first analysis, `FM` is the focal sequence and `SM` is the partner sequence.
+The univariate example contains one categorical variable for the first member (`FM`) and the second member (`SM`) of a dyad. Each row corresponds to one measurement occasion. In the first analysis, `FM` is the first member sequence analyzed and `SM` is the second member sequence.
 
 ```{r univariate-data}
 utils::data("dyadic_univariate_example", package = "dyadicMarkov")
@@ -100,9 +100,9 @@ The returned object is a list with an additional S3 class. It can therefore be p
 
 ## Interpretation
 
-In this example, the selected pattern is `PM (A3)`. This indicates that the previous state of the partner is retained in the restricted structure, whereas the previous state of the focal individual is not retained. In the terminology of the univariate method, this corresponds to a partner-only pattern.
+In this example, the selected pattern is `PM (A3)`. This indicates that the previous state of the second member is retained in the restricted structure, whereas the previous state of the first member is not retained. In the terminology of the univariate method, this corresponds to a partner-only pattern.
 
-The result should be interpreted as a pattern description for the focal sequence analyzed by the function. The function call above uses `FM` as the focal sequence and `SM` as the partner sequence. Reversing the two arguments analyzes the sequence from the perspective of the second member. Thus, describing both members of a dyad requires two calls, and each returned pattern is specific to the focal sequence.
+The result should be interpreted as a pattern description for the member sequence analyzed by the function. The function call above analyzes `FM` as the first member sequence and `SM` as the second member sequence. Reversing the two arguments analyzes the sequence from the perspective of the second member. Thus, describing both members of a dyad requires two calls, and each returned pattern is specific to the member sequence analyzed.
 
 ```{r reverse-perspective}
 pat_uni_reverse <- dyadicMarkov::univariatePattern(
