+++ vignettes/bivariate-workflow.Rmd
@@ -56,7 +56,7 @@ dim(emp_bi)
 
 ## Global bivariate case
 
-The function `bivariateCase()` performs the global step of the bivariate method. It applies the package chi-squared comparisons to classify the focal sequence dependence structure as a trivial, univariate, partial bivariate or complete bivariate case.
+The function `bivariateCase()` performs the global step of the bivariate method. It applies the package chi-squared comparisons to classify the dependence structure of the member sequence analyzed as a trivial, univariate, partial bivariate or complete bivariate case.
 
 ```{r bivariate-case}
 case_bi <- dyadicMarkov::bivariateCase(emp_bi, alpha = 0.05)
@@ -82,11 +82,11 @@ complete_bi$aic
 
 In this example, the selected complete bivariate pattern is `D2`, labelled by the package as complete actor on the main, actor-partner on the second.
 
-## Reassigning the focal and main roles
+## Reassigning member and main-variable roles
 
-The previous analysis used `FM_V1` as the focal sequence, `V1` as the main variable and `V2` as the second variable. In the bivariate method, these are analysis roles rather than fixed identities. To describe the dyad more completely, the roles can be reassigned so that each member-variable sequence is analyzed in turn.
+The previous analysis used `FM_V1` as the member-variable sequence analyzed, `V1` as the main variable and `V2` as the second variable. In the bivariate method, these are analysis roles rather than fixed identities. To describe the dyad more completely, the roles can be reassigned so that each member-variable sequence is analyzed in turn.
 
-The same data set gives different branches of the global-and-local procedure depending on the focal sequence and the main variable.
+The same data set gives different branches of the global-and-local procedure depending on the member-variable sequence analyzed and the main variable.
 
 ```{r bivariate-role-rotation}
 analyze_bivariate <- function(label, fm_v1, sm_v1, fm_v2, sm_v2) {
@@ -119,30 +119,30 @@ analyze_bivariate <- function(label, fm_v1, sm_v1, fm_v2, sm_v2) {
 d <- dyadic_bivariate_example
 
 analyze_bivariate(
-  "FM_V1 as focal sequence, V1 as main variable",
+  "FM_V1 as analyzed sequence, V1 as main variable",
   d$FM_V1, d$SM_V1, d$FM_V2, d$SM_V2
 )
 
 analyze_bivariate(
-  "SM_V1 as focal sequence, V1 as main variable",
+  "SM_V1 as analyzed sequence, V1 as main variable",
   d$SM_V1, d$FM_V1, d$SM_V2, d$FM_V2
 )
 
 analyze_bivariate(
-  "FM_V2 as focal sequence, V2 as main variable",
+  "FM_V2 as analyzed sequence, V2 as main variable",
   d$FM_V2, d$SM_V2, d$FM_V1, d$SM_V1
 )
 
 analyze_bivariate(
-  "SM_V2 as focal sequence, V2 as main variable",
+  "SM_V2 as analyzed sequence, V2 as main variable",
   d$SM_V2, d$FM_V2, d$SM_V1, d$FM_V1
 )
 ```
 
-For this example, the four role assignments illustrate three branches of the procedure: complete bivariate cases for the two `V1` focal sequences, a partial bivariate case for `FM_V2`, and a univariate case for `SM_V2`. The partial branch is therefore demonstrated with the same bivariate example data rather than with an artificial seeded example.
+For this example, the four role assignments illustrate three branches of the procedure: complete bivariate cases for the two `V1` member-variable sequences, a partial bivariate case for `FM_V2`, and a univariate case for `SM_V2`. The partial branch is therefore demonstrated with the same bivariate example data rather than with an artificial seeded example.
 
 ## Reading the global and local steps together
 
[…]
