### 2026-06-21 b92a757
Polish public documentation for CRAN release
