+++ README.md
@@ -4,7 +4,7 @@
 
 <table>
   <tr>
-    <td style="vertical-align: middle; padding-right: 14px;">Pattern Identification<br>for Dyadic Sequences<br>Using Transition Matrices</td>
+    <td style="vertical-align: middle; padding-right: 14px;">Pattern Estimation and Identification<br>for Dyadic Sequences<br>Using Transition Matrices in R</td>
     <td>
       <a href="https://CRAN.R-project.org/package=dyadicMarkov"><img src="https://www.r-pkg.org/badges/version-ago/dyadicMarkov" alt="CRAN version and release age"></a><br>
       <a href="https://cran.r-project.org/web/checks/check_results_dyadicMarkov.html"><img src="https://badges.cranchecks.info/worst/dyadicMarkov.svg" alt="CRAN checks"></a><br>
@@ -26,9 +26,23 @@
 
 ## Overview
 
-`dyadicMarkov` provides methods for analyzing dyadic interaction sequences using transition matrices within the Actor-Partner Interdependence Model. The package supports the computation of empirical transition counts, maximum likelihood estimation of transition probabilities and identification of interaction patterns in univariate and bivariate dyadic interaction sequences.
+`dyadicMarkov` provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
 
-The package is designed for categorical dyadic sequences, where two members of a dyad are observed repeatedly over time. It allows researchers to model the temporal dynamics of each member while accounting for the possible influence of their partner.
+The package is designed for settings in which two members of a dyad are observed repeatedly over time. Examples include daily diary data, coded interaction sequences, repeated binary responses, and intensive longitudinal designs in which researchers want to describe how each member's next state relates to their own previous state and to their partner's previous state.
+
+The core question addressed by the package is whether the observed transition structure is best described by an actor-partner, actor-only, partner-only, independence, partial bivariate, or complete bivariate pattern, depending on the workflow used.
+
+## When to use this package
+
+Use `dyadicMarkov` when:
+
+* two members of one dyad are observed repeatedly over ordered measurement occasions;
+* the observed states are categorical and coded as integers;
+* the analysis focuses on transition structures rather than continuous outcomes;
+* the research question concerns temporal dependence, dyadic dependence, or both;
+* the goal is to estimate transition probabilities and identify interpretable interaction patterns.
+
+The univariate workflow analyzes one categorical variable observed for two dyad members. The bivariate workflow analyzes two categorical variables observed for both members. The current bivariate implementation supports binary variables (`states = 2`). With two binary variables observed for two members, the previous state is described by four binary components, producing a bivariate empirical count matrix with 16 rows and 2 columns.
 
 ## Installation
 
@@ -38,38 +52,91 @@ You can install the released version of `dyadicMarkov` from CRAN:
 install.packages("dyadicMarkov")
 ```
 
+You can install the development version from R-universe:
+
+```r
+install.packages(
+  "dyadicMarkov",
+  repos = c("https://boellenruecherm.r-universe.dev", "https://cloud.r-project.org")
+)
+```
+
[…]
