+++ man/dyadicMarkov-package.Rd
@@ -4,8 +4,8 @@
 \name{dyadicMarkov-package}
 \alias{dyadicMarkov}
 \alias{dyadicMarkov-package}
-\title{dyadicMarkov: Pattern Identification for Dyadic Sequences Using
-Transition Matrices}
+\title{dyadicMarkov: Pattern Estimation and Identification for Dyadic Sequences Using
+Transition Matrices in R}
 \description{
 The dyadicMarkov package provides tools for analyzing categorical dyadic
 sequences using transition matrices in the Longitudinal Actor-Partner
