+++ index.md
@@ -9,10 +9,39 @@ The package supports empirical transition counts, maximum likelihood
 estimation of transition probabilities, and identification of univariate
 and bivariate patterns of interaction in dyadic sequences.
 
-The package is designed for single-case dyadic sequence analysis, where
-two members of a dyad are observed repeatedly over time. It allows
-researchers to describe the temporal dynamics of each member while
-accounting for the possible influence of their partner.
+The package is designed for settings in which two members of a dyad are
+observed repeatedly over time. Examples include daily diary data, coded
+interaction sequences, repeated binary responses, and intensive
+longitudinal designs in which researchers want to describe how each
+member’s next state relates to their own previous state and to their
+partner’s previous state.
+
+The core question addressed by the package is whether the observed
+transition structure is best described by an actor-partner, actor-only,
+partner-only, independence, partial bivariate, or complete bivariate
+pattern, depending on the workflow used.
+
+## When to use this package
+
+Use `dyadicMarkov` when:
+
+- two members of one dyad are observed repeatedly over ordered
+  measurement occasions;
+- the observed states are categorical and coded as integers;
+- the analysis focuses on transition structures rather than continuous
+  outcomes;
+- the research question concerns temporal dependence, dyadic dependence,
+  or both;
+- the goal is to estimate transition probabilities and identify
+  interpretable interaction patterns.
+
+The univariate workflow analyzes one categorical variable observed for
+two dyad members. The bivariate workflow analyzes two categorical
+variables observed for both members. The current bivariate
+implementation supports binary variables (`states = 2`). With two binary
+variables observed for two members, the previous state is described by
+four binary components, producing a bivariate empirical count matrix
+with 16 rows and 2 columns.
 
 ## Installation
 
@@ -44,25 +73,46 @@ Then load the package with:
 library(dyadicMarkov)
 ```
 
+## Example data
+
+The package includes small synthetic datasets that illustrate the
+required input structure and the package workflow. They are not intended
+to represent a substantive empirical study. They are used to show how
+ordered categorical dyadic sequences are transformed into transition
+count matrices and then into pattern-identification results.
+
+For complete examples, see the workflow vignettes:
+
+- [Introduction to
+  dyadicMarkov](https://boellenruecherm.github.io/dyadicMarkov-public/articles/dyadicMarkov-introduction.md)
+- [Univariate dyadic
+  workflow](https://boellenruecherm.github.io/dyadicMarkov-public/articles/univariate-workflow.md)
+- [Bivariate dyadic
+  workflow](https://boellenruecherm.github.io/dyadicMarkov-public/articles/bivariate-workflow.md)
+
 ## Methodological background
 
 The methodological foundation of `dyadicMarkov` is based on research on
-dyadic sequences, the Longitudinal Actor-Partner Interdependence Model,
-and Markov chains.
+categorical dyadic sequences, the Longitudinal Actor-Partner
+Interdependence Model, and Markov chains.
 
 The 2023 article introduces the use of Markov chains for modeling
 categorical longitudinal dyadic data in a single-case perspective. The
 2024 article extends the approach to the identification, visualization,
 and clustering of similar behaviors in samples of dyads. The bivariate
-extension is documented in related work by Bollenrücher, Antonietti, and
-Darwiche.
+extension is documented in related work by Bollenrücher, Darwiche, and
+Antonietti.
+
+The visualization and clustering methodology is part of the
[…]
