### 2026-06-21 a9524c7
Clarify vignette use cases for rOpenSci
