+++ vignettes/dyadicMarkov-introduction.Rmd
@@ -17,6 +17,14 @@ knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 
 `dyadicMarkov` is based on three methodological papers on dyadic pattern analysis with the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov chains. The univariate single-case method is described by Bollenrücher, Darwiche, and Antonietti (2023). The extension to visualization and clustering of similar dyadic behaviors is described by Bollenrücher, Darwiche, and Antonietti (2024). The bivariate single-case method is described by Böllenrücher, Darwiche, and Antonietti (in press). The visualization and clustering methodology of Bollenrücher, Darwiche, and Antonietti (2024) is methodological background; it is not currently part of the exported package API.
 
+## Real-world use cases
+
+The package is intended for ordered categorical observations collected from two members of a dyad. In practice, such sequences may arise from coded interaction data, daily diary studies, repeated binary responses, or intensive longitudinal designs. For example, researchers may code whether each partner shows a given behavior at each measurement occasion, whether a parent and child are in one of several interaction states, or whether two individuals report the presence or absence of a response across repeated observations.
+
+The package does not estimate a general-purpose time-series forecasting model. Instead, it summarizes how the next state of the member sequence analyzed is associated with its own previous state and with the previous state of the partner. The resulting pattern labels help describe whether the observed transitions are better characterized by actor dependence, partner dependence, actor-partner dependence, independence, or, in the bivariate workflow, by partial or complete bivariate dependence structures.
+
+The example datasets included in the package are synthetic. They are used to make the required input structure reproducible and easy to inspect. They should be read as small stand-ins for real ordered dyadic sequences, not as substantive empirical datasets.
+
 ## Data structure
 
 The package works with categorical dyadic sequences. In the univariate case, one categorical variable is observed over time for two members of a dyad. For each function call, the first member is the member whose next state is modeled, and the second member supplies the partner sequence. The roles can be reversed to analyze the other member.
