### 2026-06-21 44caa28
Update license information to specify MIT License
