### 2026-07-14 b330638
Update CRAN checks badge in README.md
