### 2026-07-14 1c6a5ab
Update CRAN checks badge in README.md
