### 2026-08-02 b6e3e00
Improve bivariate documentation and math rendering
