+++ _pkgdown.yml
@@ -2,6 +2,7 @@ url: https://boellenruecherm.github.io/dyadicMarkov-public/
 
 template:
   bootstrap: 5
+  math-rendering: mathjax
 
 navbar:
   structure:
