### 2026-08-21 0aff4cd
Sync public repo with dyadicMarkov 0.1.2
