+++ R/apim-bivariate.R
@@ -25,7 +25,16 @@
 }
 
 
-# Validate that bivariate counting currently uses binary states.
+#' Validate binary bivariate state count
+#'
+#' Internal helper that validates the current binary-state requirement for
+#' bivariate functions.
+#'
+#' @param states Number of categorical states.
+#'
+#' @return The integer value `2L` when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_states <- function(states) {
   ok_states <- !missing(states) &&
     is.numeric(states) &&
@@ -45,7 +54,15 @@
 }
 
 
-# Validate that all bivariate chains are numeric.
+#' Validate bivariate chain types
+#'
+#' Internal helper that checks whether all bivariate chain inputs are numeric.
+#'
+#' @param chains List of bivariate chain vectors.
+#'
+#' @return Invisibly returns `chains` when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_chain_types <- function(chains) {
   is_num <- vapply(chains, is.numeric, logical(1))
 
@@ -57,7 +74,16 @@
 }
 
 
-# Validate that bivariate chains do not contain missing values.
+#' Validate missing values in bivariate chains
+#'
+#' Internal helper that checks whether bivariate chain inputs contain missing
+#' values.
+#'
+#' @param chains List of bivariate chain vectors.
+#'
+#' @return Invisibly returns `chains` when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_chain_missing <- function(chains) {
   has_na <- vapply(chains, anyNA, logical(1))
 
@@ -69,7 +95,15 @@
 }
 
 
-# Validate that bivariate chains contain finite numeric values.
+#' Validate finite values in bivariate chains
+#'
+#' Internal helper that checks whether all bivariate chain values are finite.
+#'
+#' @param chains List of bivariate chain vectors.
+#'
+#' @return Invisibly returns `chains` when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_chain_finite <- function(chains) {
   finite <- vapply(
     chains,
@@ -88,7 +122,16 @@
 }
 
 
-# Validate that all bivariate chains have a common usable length.
+#' Validate bivariate chain lengths
+#'
+#' Internal helper that checks whether all bivariate chains have the same
+#' length and contain at least two observations.
+#'
+#' @param chains List of bivariate chain vectors.
+#'
+#' @return The common chain length when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_chain_lengths <- function(chains) {
   lengths <- vapply(chains, length, integer(1))
 
@@ -106,7 +149,16 @@
 }
 
 
-# Validate that bivariate chains use integer-coded states in {1, 2}.
+#' Validate bivariate chain values
+#'
+#' Internal helper that checks whether bivariate chain values are integer coded
+#' and belong to the supported binary state set \code{1, 2}.
+#'
+#' @param chains List of bivariate chain vectors.
+#'
+#' @return Invisibly returns `chains` when validation succeeds; otherwise raises
+#'   an error.
+#' @noRd
 .validate_bivariate_chain_values <- function(chains) {
   bad_chain <- function(x) {
     any(x != as.integer(x)) || any(x < 1L | x > 2L)
@@ -212,19 +264,27 @@ countEmpBivariate <- function(
 #' Bivariate case identification for dyadic Markov chains
 #'
 #' Identifies the bivariate case as \code{"trivial"}, \code{"univariate"},
-#' \code{"partial"}, or \code{"complete"} using two likelihood-ratio tests
-#' against constrained bivariate structures.
[…]
