+++ articles/bivariate-workflow.md
@@ -8,6 +8,13 @@ observed repeatedly for the two members of a dyad. The bivariate method
 follows the global-and-local procedure described in Böllenrücher et al.
 (in press).
 
+The bivariate method uses matrix codes to identify the local dependence
+patterns. Partial bivariate patterns are denoted B1–B3, while complete
+bivariate patterns are denoted C, D1–D4, and E1–E4. When the global step
+identifies a univariate case, the A-family codes described in the
+univariate workflow apply. The pattern nomenclature is summarized in
+Table 2 of Böllenrücher et al. (in press).
+
 The current bivariate functions support `states = 2`. With two binary
 variables observed for two members, the previous state is described by
 four binary components. The empirical bivariate count matrix therefore
@@ -25,8 +32,8 @@ The 16 rows arise because the previous state combines four binary lagged
 components: first member on `V1`, second member on `V1`, first member on
 `V2`, and second member on `V2`. With two possible states for each
 component, this gives \\2^4 = 16\\ previous-state combinations. The 2
-columns represent the possible next states of the member sequence
-analyzed on the current main variable.
+columns represent the possible next states of the first member on the
+current main variable.
 
 ## Data
 
@@ -112,10 +119,14 @@ dim(emp_bi)
 
 The function
 [`bivariateCase()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/bivariateCase.md)
-performs the global step of the bivariate method. It applies the package
-chi-squared comparisons to classify the dependence structure of the
-member sequence analyzed as a trivial, univariate, partial bivariate or
-complete bivariate case.
+performs the global step of the bivariate method. The global approach
+compares nested models within the likelihood-ratio test (LRT) framework.
+[`bivariateCase()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/bivariateCase.md)
+implements the two chi-squared tests for the A1 and B1 comparisons used
+to identify the dependence case. `dyadicMarkov` evaluates these tests
+using Pearson’s chi-squared statistic, \\X^2 = \sum (O-E)^2/E\\, to
+classify the analyzed sequence as a trivial, univariate, partial
+bivariate, or complete bivariate case.
 
 ``` r
 
@@ -149,8 +160,9 @@ patterns.
 
 For a complete bivariate case,
 [`completePattern()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/completePattern.md)
-compares the complete bivariate candidate structures by AIC and returns
-the selected pattern.
+first computes the G-squared deviance, \\G^2 = 2\sum O\log(O/E)\\, for
+each complete bivariate candidate structure. It then calculates \\AIC =
+G^2 + 2k\\ and returns the candidate with the smallest AIC.
 
 ``` r
 
@@ -158,47 +170,42 @@ complete_bi <- dyadicMarkov::completePattern(emp_bi)
 
 complete_bi
 #> Dyadic interaction pattern
-#> Pattern: complete actor on the main, actor partner on the second (D2)
+#> Pattern: actor only on the main, actor-partner on the second (D2)
 complete_bi$pattern
-#> [1] "complete actor on the main, actor partner on the second (D2)"
+#> [1] "actor only on the main, actor-partner on the second (D2)"
 complete_bi$aic
-#>                                                          pattern matrix
-#> 1                             complete actor partner on both (C)      C
-#> 2 complete partner on the main, actor partner on the second (D1)     D1
-#> 3   complete actor on the main, actor partner on the second (D2)     D2
-#> 4 complete actor partner on the main, partner on the second (D3)     D3
-#> 5   complete actor partner on the main, actor on the second (D4)     D4
-#> 6                                  complete partner on both (E1)     E1
[…]
