+++ reference/dyadicMarkov-package.md
@@ -14,12 +14,34 @@ The temporal structure is represented by the order of the observations.
 The main outputs are empirical count matrices, estimated transition
 probability matrices, and identified patterns of interaction.
 
+## Supported state spaces
+
+The univariate workflow supports any integer number of categorical
+states \\S \ge 2\\. The bivariate method is developed for two
+dichotomous variables and therefore supports `states = 2` only,
+producing 16-by-2 empirical count matrices. Bivariate support beyond two
+states would require additional mathematical and software development
+and is not implemented by this package.
+
+## Comparison statistics
+
+The univariate pattern-identification procedure is a likelihood-ratio
+test (LRT) procedure that dyadicMarkov evaluates using Pearson's
+chi-squared statistic. The global bivariate approach compares nested
+models within an LRT framework;
+[`bivariateCase()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/bivariateCase.md)
+implements two chi-squared tests for the A1 and B1 comparisons,
+evaluated using Pearson's chi-squared statistic. Local partial and
+complete bivariate pattern selection is separate: it computes the
+G-squared deviance and then \\AIC = G^2 + 2k\\ for each candidate
+structure.
+
 ## Main terminology
 
 A dyadic sequence records the categorical states of two interacting
 individuals over time. Empirical transition counts summarize transitions
-from previous dyadic states to subsequent states of the member sequence
-analyzed. Transition probabilities are estimated by normalizing each row
+from previous dyadic states to subsequent states of the analyzed
+sequence. Transition probabilities are estimated by normalizing each row
 of the empirical transition count matrix. Patterns of interaction are
 identified by comparing unrestricted and restricted transition
 structures based on actor and partner effects.
@@ -35,7 +57,7 @@ functions are intended to remain stable across minor releases.
 The method models categorical dyadic sequences with Markov chains in the
 Longitudinal Actor-Partner Interdependence Model (L-APIM) framework. It
 uses transition matrices to represent how previous dyadic states are
-related to the current state of the member sequence analyzed.
+related to the current state of the analyzed sequence.
 
 ## Algorithmic contribution
 
@@ -84,11 +106,13 @@ Useful links:
 
 ## Author
 
-**Maintainer**: Mattia Böllenrücher
-<mattia.boellenruecher@student.unisg.ch>
+**Maintainer**: Mattia Böllenrücher <mboellenruec@student.ethz.ch>
 ([ORCID](https://orcid.org/0009-0004-4149-4745)) \[copyright holder\]
 
 Authors:
+
+- Mattia Böllenrücher <mboellenruec@student.ethz.ch>
+  ([ORCID](https://orcid.org/0009-0004-4149-4745)) \[copyright holder\]
 
 - Mégane Bollenrücher ([ORCID](https://orcid.org/0000-0002-9035-8799))
 
