+++ authors.md
@@ -18,13 +18,13 @@ Source:
 
 Böllenrücher M, Bollenrücher M, Antonietti J (2026). *dyadicMarkov:
 Pattern Estimation and Identification for Dyadic Sequences Using
-Transition Matrices in R*. R package version 0.1.1,
+Transition Matrices in R*. R package version 0.1.2,
 <https://github.com/BoellenruecherM/dyadicMarkov-public>.
 
     @Manual{,
       title = {dyadicMarkov: Pattern Estimation and Identification for Dyadic Sequences Using Transition Matrices in R},
       author = {Mattia Böllenrücher and Mégane Bollenrücher and Jean-Philippe Antonietti},
       year = {2026},
-      note = {R package version 0.1.1},
+      note = {R package version 0.1.2},
       url = {https://github.com/BoellenruecherM/dyadicMarkov-public},
     }

[…]
