### 2026-08-21 baa5780
Polish public repository and pkgdown site
