+++ NEWS.md
@@ -1,4 +1,4 @@
-# dyadicMarkov 0.1.2
+# dyadicMarkov 0.1.2 (2026-08-21)
 
 * Clarified that the univariate pattern-identification procedure is an LRT
   procedure evaluated using Pearson's chi-squared statistic, while the global bivariate
@@ -17,7 +17,7 @@
 * Made the manual simulated-parity script stop when a comparison fails.
 * Declared `srr` as a development/documentation dependency.
 
-# dyadicMarkov 0.1.1
+# dyadicMarkov 0.1.1 (2026-06-21)
 
 * Updated package wording and metadata for the CRAN submission.
 * Added S3 classes and print/summary support for pattern and case identification results.
@@ -33,6 +33,6 @@
   behavior.
 * Improved bivariate count validation coverage for unsupported and malformed inputs.
 
-# dyadicMarkov 0.1.0
+# dyadicMarkov 0.1.0 (2026-03-16)
 
 * Initial CRAN submission.
