+++ README.md
@@ -1,6 +1,5 @@
 # dyadicMarkov
 
-<!-- cranlogs:start -->
 
 <table>
   <tr>
@@ -20,10 +19,6 @@
 
 <sub>Download counts are recorded from the RStudio/Posit CRAN mirror via <code>cranlogs</code>.</sub>
 
-<!-- Last automatic update: 2026-05-25 -->
-
-<!-- cranlogs:end -->
-
 ## Overview
 
 `dyadicMarkov` provides methods for analyzing categorical dyadic sequences using transition matrices within the Longitudinal Actor-Partner Interdependence Model (L-APIM) and Markov-chain framework. The package supports empirical transition counts, maximum likelihood estimation of transition probabilities, and identification of univariate and bivariate patterns of interaction in dyadic sequences.
@@ -42,7 +37,7 @@ Use `dyadicMarkov` when:
 * the research question concerns temporal dependence, dyadic dependence, or both;
 * the goal is to estimate transition probabilities and identify interpretable interaction patterns.
 
-The univariate workflow analyzes one categorical variable observed for two dyad members. The bivariate workflow analyzes two categorical variables observed for both members. The current bivariate implementation supports binary variables (`states = 2`). With two binary variables observed for two members, the previous state is described by four binary components, producing a bivariate empirical count matrix with 16 rows and 2 columns.
+The univariate workflow analyzes one categorical variable observed for two dyad members and supports any integer number of categorical states (`states >= 2`). The bivariate workflow analyzes two categorical variables observed for both members. The current bivariate implementation supports binary variables (`states = 2`) only. With two binary variables observed for two members, the previous state is described by four binary components, producing a bivariate empirical count matrix with 16 rows and 2 columns.
 
 ## Installation
 
@@ -149,3 +144,7 @@ citation("dyadicMarkov")
 ## License
 
 This package is licensed under the MIT License. See [LICENSE.md](LICENSE.md) for details.
+
+## Code of Conduct
+
+Please note that this project follows the [rOpenSci Code of Conduct](https://ropensci.org/code-of-conduct/). By contributing to this project, you agree to abide by its terms.
