### 2026-08-21 #773 @BoellenruecherM
Dear Mark,

Thank you again for your detailed feedback on the pre-submission inquiry. We have now completed a substantial revision of `dyadicMarkov` addressing the points raised above, and version 0.1.2 has been released on CRAN.

In particular, we have worked on the following aspects.

### Statistical standards and package category

The package is now identified by `pkgcheck` as belonging to the **Exploratory Data Analysis** category. We have reviewed and documented the applicable rOpenSci statistical standards for that category in addition to the General standards.

A fresh `srr::srr_stats_pre_submit()` check reports:
All applicable standards [v0.2.0] have been documented in this package
(53 complied with; 49 N/A standards)

We tried to apply the EDA standards wherever they are relevant to the package rather than marking standards N/A simply to obtain formal compliance.

### General use case and documentation

We substantially expanded the documentation to make the intended scope and general use of the package clearer. The README and pkgdown website now explain that `dyadicMarkov` is intended for categorical longitudinal dyadic sequences in which two members of a dyad are repeatedly observed over time, with examples including daily diary data, coded interaction sequences, repeated binary responses, and intensive longitudinal designs.

The package now contains three vignettes: an introduction, a dedicated univariate workflow, and a dedicated bivariate workflow. These provide reproducible examples using package datasets and explain the inputs, transition-matrix construction, estimation, pattern identification, and interpretation of the results. All exported functions also have examples.

We also clarified the distinction between the two methodological scopes. The **univariate procedure is not restricted to three states** and supports any integer number of categorical states (`states >= 2`). The current **bivariate methodology is developed and validated for two dichotomous variables (`states = 2`)**. We therefore retained that restriction rather than generalizing the implementation beyond the methodology supported by the underlying work. For two binary variables observed for both dyad members, the previous state consists of four binary components, giving the 16-row bivariate empirical count matrix discussed previously.

### Testing and validation

Test coverage has increased substantially and is currently reported by `pkgcheck` as **98.2%**.

We also performed the final checks from a fresh clone of the public repository to avoid any influence from local development metadata. On R 4.6.1 under Windows 11:
R CMD check --as-cran
0 errors | 0 warnings | 0 notes

A fresh `pkgcheck` run additionally reports:
✔ Package coverage is 98.2%.
✔ This is a statistical package which complies with all applicable standards
✔ R CMD check found no errors.
✔ R CMD check found no warnings.

`pkgcheck` locally reports that it does not detect continuous integration checks, although the public repository does have GitHub Actions CI and the current workflows are running successfully.

### Current release

The current version is `dyadicMarkov` 0.1.2, now available on CRAN. The public repository and pkgdown website have also been updated to correspond to this release and the revisions above.

We hope these changes address the concerns about statistical-review scope, the general use case, documentation, and the previously unexplained restrictions. In particular, we have tried to improve the package without extending the statistical methodology beyond what has actually been developed and validated.

Would you be willing to reassess the package in light of these changes and let us know whether it can now proceed from the pre-submission inquiry to formal rOpenSci statistical software review?

Thank you again for your time and guidance.

Best regards,
Mattia
