### 2026-08-24 #773 @mpadge
Ah, yes, @BoellenruecherM, the changes you've made have indeed enormously improved my understanding of the package. It's now much clearer, and I think we're good to proceed to a full submission.

Before we do that, however, I think you could likely increase your compliance with our standards. The `srr` standards are a form of recommendation of what good stats packages should do. In your case, one key aspect of the EDA standards that is missing is a `plot` method. Implementing this would great enhance the usability of your package, and would also be the single best way to increase compliance with standards. These are some of the standards currently not complied with yet for which a `plot` method would easily cover:

## EDA

**Return Results / Output Data**
- `EA4.2` (plot half) — No `plot` method exists for any S3 class (`dyadic_pattern`, `dyadic_case`, `dyadic_counts`, `dyadic_mle`); print/summary are covered, plot isn't.

**Visualization and Summary Output**
- `EA5.0` / `EA5.0a` / `EA5.0b` — No accessibility-oriented typeface/colour defaults, since no plots exist yet.
- `EA5.1` — No documented accessibility consideration for typeface overrides.
- `EA5.4` — No visualisation exists yet to apply sensible value rounding (e.g. `pretty()`) to.
- `EA5.5` — No visualisation exists yet to carry axis units.

**Testing**
- `EA6.0d` — The `aic` data frame returned by `partialPattern()`/`completePattern()` isn't tested for column classes (character `pattern`/`matrix`, numeric `aic`).
- `EA6.1` — No graphical-output tests (e.g. `vdiffr`), contingent on plot methods existing.

## General

**Documentation**
- `G1.5` — No script/vignette reproducing the performance claims cited from the source papers (e.g. the `@bollen2026` sequence-length sensitivity analysis).

**Input Structures**
- `G2.0a` — `@param` docs don't state that `chainFM`/`chainSM` must be equal length (only the runtime error does).
- `G2.1a` — `@param` docs don't explicitly state chain vectors must be numeric/integer-coded (only the runtime error does).

**Testing**
- `G5.7` — No test showing pattern-classification accuracy improves/stabilizes as chain length increases (the vignette claims this, but it remains untested).
- `G5.9a` — No test showing trivial numeric noise on empirical count matrices leaves `mleEstimation()` results unchanged.

You may also consider implementing extended tests, generally run in local-only environments. This may then cover standards **G5.108** and **G5.12**.

---

## Conclusion

We invite you to close this issue, and open a full submission. And we also suggest that you first try to expand compliance with `srr` standards, particularly through implementing `plot` methods for (at least some of) your S3-class objects. Thanks!
