### 2026-08-24 #773 @BoellenruecherM
Thank you very much, Mark. This is extremely helpful. In particular, your suggestion regarding plot methods makes a lot of sense for the package.

We will therefore take some additional time before opening the full submission to investigate and implement scientifically appropriate visualizations for the package's S3 objects. We will also work through the other standards you identified.

Thank you again for the detailed guidance and for helping us improve the package substantially during the pre-submission process.
