### 2026-08-24 #773 @BoellenruecherM
Given how useful these additional suggestions are, I would actually prefer to take a little more time to work through them carefully before moving to the full submission.

This could take a few days or weeks, as I would like to make sure that any plots we introduce are scientifically consistent with the methodology.

If it is convenient for you and @adamhsparks, would it perhaps make sense to leave this pre-submission open and place it on Holding while I work through these improvements?
