+++ tests/testthat/test-s3-summary-tables.R
@@ -1,0 +1,1017 @@
+#' @srrstats {EA6.0c} Summary-table tests verify stable column names and source-faithful column and candidate ordering for univariate tests, bivariate global tests, and partial and complete AIC comparisons.
+#' @srrstats {EA6.0d} Summary-table tests verify expected dimensions and column types for restriction/model labels, statistics, degrees of freedom, p-values, decisions, AIC values, Delta AIC values, minimum indicators, and selected indicators.
+#' @srrstats {EA6.0e} Summary-table tests verify inferential and derived values, including p-value decisions, Delta AIC calculations, selected-pattern consistency, numerical AIC ties, and non-finite AIC handling.
+#' @noRd
+NULL
+make_bivariate_summary_example <- function() {
+
+  chainFM_V1 <- c(1L, 2L, 1L, 2L, 2L, 1L)
+  chainSM_V1 <- c(2L, 1L, 2L, 1L, 1L, 2L)
+  chainFM_V2 <- c(1L, 1L, 2L, 2L, 1L, 2L)
+  chainSM_V2 <- c(2L, 2L, 1L, 1L, 2L, 1L)
+
+  dyadicMarkov::countEmpBivariate(
+    chainFM_V1,
+    chainSM_V1,
+    chainFM_V2,
+    chainSM_V2,
+    states = 2L
+  )
+}
+
+
+expected_decision <- function(p_value, alpha) {
+
+  valid <-
+    is.finite(p_value) &
+    p_value >= 0 &
+    p_value <= 1
+
+  out <- rep(NA_character_, length(p_value))
+
+  out[valid] <- ifelse(
+    p_value[valid] <= alpha,
+    "Rejected",
+    "Not rejected"
+  )
+
+  out
+}
+
+
+expected_aic_augmentation <- function(aic) {
+
+  finite <- is.finite(aic)
+
+  delta <- aic
+  minimum <- rep(FALSE, length(aic))
+
+  if (any(finite)) {
+
+    min_aic <- min(aic[finite])
+
+    delta[finite] <-
+      aic[finite] - min_aic
+
+    tolerance <-
+      sqrt(.Machine$double.eps) *
+      max(
+        1,
+        abs(min_aic)
+      )
+
+    minimum[finite] <-
+      abs(
+        aic[finite] -
+          min_aic
+      ) <= tolerance
+
+  } else {
+
+    delta[] <- NA_real_
+  }
+
+  list(
+    delta = delta,
+    minimum = minimum
+  )
+}
+
+
+test_that("univariate summary table preserves inferential evidence", {
+
+  chainFM <- c(1L, 2L, 1L, 2L, 2L, 1L)
+  chainSM <- c(2L, 1L, 2L, 1L, 1L, 2L)
+
+  alpha <- 0.05
+
+  object <- dyadicMarkov::univariatePattern(
+    chainFM,
+    chainSM,
+    states = 2L,
+    alpha = alpha
+  )
+
+  out <- summary(object)
+
+  expect_s3_class(
+    out,
+    "summary_dyadic_pattern"
+  )
+
+  expect_named(
+    out,
+    c(
+      "pattern",
+      "tests",
+      "alpha",
+      "states",
+      "call"
+    )
+  )
+
+  expect_identical(
+    names(out$tests),
+    c(
+      "restriction",
+      "statistic",
+      "df",
+      "p_value",
+      "decision"
+    )
+  )
+
+  expect_identical(
+    nrow(out$tests),
+    2L
+  )
+
+  expect_identical(
+    out$tests$restriction,
+    c(
+      "AM (A2): actor-only restriction",
+      "PM (A3): partner-only restriction"
+    )
+  )
+
+  expect_type(
+    out$tests$restriction,
[…]
