+++ R/plot-methods.R
@@ -1,0 +1,887 @@
+#' Validate plot text scaling
+#'
+#' @param cex Positive finite scalar or `NULL`.
+#'
+#' @return Numeric scalar or `NULL`.
+#' @noRd
+.validate_dyadic_plot_cex <- function(cex) {
+  if (is.null(cex)) {
+    return(NULL)
+  }
+
+  if (
+    !is.numeric(cex) ||
+    length(cex) != 1L ||
+    is.na(cex) ||
+    !is.finite(cex) ||
+    cex <= 0
+  ) {
+    stop(
+      "`cex` must be one positive finite number or `NULL`.",
+      call. = FALSE
+    )
+  }
+
+  as.numeric(cex)
+}
+
+
+#' Validate plot title
+#'
+#' @param main Character scalar or `NULL`.
+#'
+#' @return Character scalar or `NULL`.
+#' @noRd
+.validate_dyadic_plot_main <- function(main) {
+  if (is.null(main)) {
+    return(NULL)
+  }
+
+  if (
+    !is.character(main) ||
+    length(main) != 1L ||
+    is.na(main)
+  ) {
+    stop(
+      "`main` must be one character string or `NULL`.",
+      call. = FALSE
+    )
+  }
+
+  main
+}
+
+
+#' Construct accessible default state colours
+#'
+#' @param states Number of categorical states.
+#'
+#' @return Character vector with one colour per state.
+#' @noRd
+.dyadic_default_state_col <- function(states) {
+  if (states == 2L) {
+    return(c("#0072B2", "#E69F00"))
+  }
+
+  grDevices::hcl.colors(
+    n = states,
+    palette = "Dark 3"
+  )
+}
+
+
+#' Validate plotting colours
+#'
+#' @param col State-colour vector or `NULL`.
+#' @param states Number of categorical states.
+#'
+#' @return Character vector with one distinct, non-fully-transparent colour
+#'   per state.
+#' @noRd
+.validate_dyadic_plot_col <- function(col, states) {
+  if (is.null(col)) {
+    col <- .dyadic_default_state_col(states)
+  }
+
+  if (
+    !is.character(col) ||
+    length(col) != states ||
+    anyNA(col)
+  ) {
+    stop(
+      "`col` must contain exactly one valid colour per state.",
+      call. = FALSE
+    )
+  }
+
+  rgba <- tryCatch(
+    grDevices::col2rgb(col, alpha = TRUE),
+    error = function(e) NULL
+  )
+
+  if (is.null(rgba)) {
+    stop(
+      "`col` contains an invalid colour.",
+      call. = FALSE
+    )
+  }
+
+  if (any(rgba[4L, ] == 0L)) {
+    stop(
+      "`col` must not contain fully transparent colours.",
+      call. = FALSE
+    )
+  }
+
+  encoded <- apply(
+    rgba,
+    2L,
+    paste,
+    collapse = ","
+  )
+
+  if (anyDuplicated(encoded)) {
+    stop(
+      "`col` must use a distinct colour for every state.",
+      call. = FALSE
+    )
+  }
+
+  col
+}
+
+
+#' Reject unsupported graphical arguments
+#'
+#' The methods retain `...` for S3 compatibility. Additional graphical
+#' arguments are deliberately not silently ignored.
+#'
+#' @param ... Additional arguments.
+#'
+#' @return Invisibly `NULL`.
+#' @noRd
+.validate_dyadic_plot_dots <- function(...) {
+  dots <- list(...)
+
+  if (length(dots) > 0L) {
+    supplied <- names(dots)
+
+    if (is.null(supplied) || !all(nzchar(supplied))) {
+      detail <- "unnamed graphical arguments"
[…]
