### 2026-09-15 7d13a08
Update pkgdown index for 0.1.3
