+++ articles/bivariate-workflow.md
@@ -15,25 +15,18 @@ identifies a univariate case, the A-family codes described in the
 univariate workflow apply. The pattern nomenclature is summarized in
 Table 2 of Böllenrücher et al. (in press).
 
-The current bivariate functions support `states = 2`. With two binary
-variables observed for two members, the previous state is described by
-four binary components. The empirical bivariate count matrix therefore
-has 16 rows and 2 columns.
+The current bivariate workflow supports binary variables (`states = 2`).
+
+The example uses the data set `dyadic_bivariate_example` included in the
+package. The data are synthetic and are used only to illustrate the
+required input structure and the package workflow.
 
 Although the data are synthetic, the four columns can be read like real
 repeated observations from a dyad. For example, `V1` could represent one
-coded behavior or response, and `V2` could represent a second coded
-behavior or response observed at the same measurement occasions. The
-columns `FM_V1` and `SM_V1` then describe the two members on the main
-variable, while `FM_V2` and `SM_V2` describe the same two members on the
-second variable.
-
-The 16 rows arise because the previous state combines four binary lagged
-components: first member on `V1`, second member on `V1`, first member on
-`V2`, and second member on `V2`. With two possible states for each
-component, this gives \\2^4 = 16\\ previous-state combinations. The 2
-columns represent the possible next states of the first member on the
-current main variable.
+coded behavior or response and `V2` a second behavior or response
+observed at the same measurement occasions. The columns `FM_V1` and
+`SM_V1` describe the two members on the main variable, while `FM_V2` and
+`SM_V2` describe the same two members on the second variable.
 
 ## Data
 
@@ -70,16 +63,36 @@ table(dyadic_bivariate_example$SM_V1)
 #> 
 #>  1  2 
 #> 18 72
+table(dyadic_bivariate_example$FM_V2)
+#> 
+#>  1  2 
+#> 21 69
+table(dyadic_bivariate_example$SM_V2)
+#> 
+#>  1  2 
+#> 20 70
 ```
 
 ## Empirical bivariate transition counts
 
-The first step is to construct the empirical bivariate transition count
-matrix with
-[`countEmpBivariate()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmpBivariate.md).
-Rows represent the 16 possible previous-state combinations of the two
-dyadic variables. Columns represent the next state of the first member
-on the main variable.
+[`countEmpBivariate()`](https://boellenruecherm.github.io/dyadicMarkov-public/reference/countEmpBivariate.md)
+computes the empirical transition counts for the first member on the
+main variable from the four observed sequences.
+
+For `states = 2`, the resulting matrix has 16 rows corresponding to the
+possible previous-state combinations of both members on both variables.
+With four binary lagged components, there are \\2^4 = 16\\ such
+combinations. The two columns correspond to the possible next states of
+the first member on the main variable.
+
+The returned `dyadic_counts` object retains ordinary matrix behavior and
+provides [`print()`](https://rdrr.io/r/base/print.html),
+[`summary()`](https://rdrr.io/r/base/summary.html), and
+[`utils::toLatex()`](https://rdrr.io/r/utils/toLatex.html) methods. The
+[`summary()`](https://rdrr.io/r/base/summary.html) method reports
[…]
