+++ articles/index.md
@@ -8,3 +8,5 @@
   workflow](https://boellenruecherm.github.io/dyadicMarkov-public/articles/univariate-workflow.md):
 - [Bivariate dyadic
   workflow](https://boellenruecherm.github.io/dyadicMarkov-public/articles/bivariate-workflow.md):
+- [Sensitivity
+  analysis](https://boellenruecherm.github.io/dyadicMarkov-public/articles/sensitivity-analysis.md):

[…]
