### 2026-09-15 52924a5
Document extended sensitivity test workflow
