### 2026-09-15 ab89bbb
Add PDF link for bivariate article
