+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   dyadicMarkov-introduction: dyadicMarkov-introduction.html
   sensitivity-analysis: sensitivity-analysis.html
   univariate-workflow: univariate-workflow.html
-last_built: 2026-09-15T13:50Z
+last_built: 2026-09-15T14:11Z
 urls:
   reference: https://boellenruecherm.github.io/dyadicMarkov-public/reference
   article: https://boellenruecherm.github.io/dyadicMarkov-public/articles
