### 2026-09-15 4ebdce1
Link sensitivity vignette from pkgdown homepage
