### 2026-09-16 07fea8f
Refine documentation for rOpenSci review
